Added patch for inhernt stats to apply for feat prerequisisites
