+++ TabletopTweaks/TabletopTweaks.csproj
@@ -186,7 +186,7 @@
     <Compile Include="Bugfixes\General\CriticalConfirmOn20.cs" />
     <Compile Include="Bugfixes\General\TemporaryHitPointDisplayFix.cs" />
     <Compile Include="MechanicsChanges\DRRework.cs" />
-    <Compile Include="MechanicsChanges\InherentSkillPoints.cs" />
+    <Compile Include="MechanicsChanges\InherentBonuses.cs" />
     <Compile Include="MechanicsChanges\MetamagicDamage.cs" />
     <Compile Include="Bugfixes\General\BlueprintProgressionFix.cs" />
     <Compile Include="MechanicsChanges\SelectiveMetamagic.cs" />
